+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/util/UploaderAnalysis.java
@@ -472,6 +472,10 @@ public class UploaderAnalysis {
   }
 
   public List<String> getFileNamesForAssetType(AssetType assetType) {
+    if (assetType.equals(AssetType.FP_MODEL_PACK)) {
+      return getFpModelPacks();
+    }
+
     List<String> result = new ArrayList<>();
     for (String file : getFilteredFilenamesWithPath()) {
       String fileName = getFileName(file);
@@ -482,14 +486,15 @@ public class UploaderAnalysis {
     return result;
   }
 
-  public String getFileNameForExtension(String extension) {
+  public List<String> getFileNamesForExtension(String extension) {
+    List<String> result = new ArrayList<>();
     for (String file : getFilteredFilenamesWithPath()) {
       String fileName = getFileName(file);
       if (fileName.toLowerCase().endsWith("." + extension.toLowerCase())) {
-        return fileName;
+        result.add(fileName);
       }
     }
-    return null;
+    return result;
   }
 
   public String getFileNameWithPathForExtension(String extension) {
@@ -528,7 +533,7 @@ public class UploaderAnalysis {
         return "This archive does not not contain a .res file.";
       }
       case ROM: {
-        if (isRom() || hasFileWithSuffixAndNot("zip", "pup", "pov")) {
+        if ((isRom() || hasFileWithSuffixAndNot("zip", "pup", "pov")) && !isFpTable()) {
           return null;
         }
         return "This archive does not not contain a ROM file.";
@@ -538,6 +543,12 @@ public class UploaderAnalysis {
           return null;
         }
         return "This archive does not not contain a DMD bundle.";
+      }
+      case FP_MODEL_PACK: {
+        if (!getFpModelPacks().isEmpty()) {
+          return null;
+        }
+        return "This archive does not not contain a Future Pinball Model bundle.";
       }
       case DIF: {
         if (hasFileWithSuffix("dif")) {
@@ -617,6 +628,7 @@ public class UploaderAnalysis {
         return "This archive does not not contain a .ini file.";
       }
       default: {
+        LOG.error("Unmapped asset type: {}", assetType + "/" + assetType.name());
         throw new UnsupportedOperationException("Unmapped asset type: " + assetType + "/" + assetType.name());
       }
     }
@@ -797,6 +809,24 @@ public class UploaderAnalysis {
     return false;
   }
 
+  private List<String> getFpModelPacks() {
+    List<String> result = new ArrayList<>();
+    for (String fileName : getFilteredFilenamesWithPath()) {
+      String suffix = FilenameUtils.getExtension(fileName);
+      String name = FilenameUtils.getBaseName(fileName);
+      if (AssetType.FPT.name().equalsIgnoreCase(suffix)) {
+        String modelFile = name + ".zip";
+        for (String s : getFilteredFilenamesWithPath()) {
+          if (s.endsWith(modelFile)) {
+            result.add(s);
+            break;
+          }
+        }
+      }
+    }
+    return result;
+  }
+
   public boolean isMusic() {
     return !isAltSound() && getMusicFolder() != null;
   }
@@ -905,6 +935,10 @@ public class UploaderAnalysis {
   }
 
   private boolean isRom() {
+    if (isFpTable()) {
+      return false;
+    }
+
     if (getFilteredFolders().isEmpty()) {
       for (String fileName : getFilteredFilenamesWithPath()) {
         String suffix = FilenameUtils.getExtension(fileName);